+++ src/debug.ts
@@ -1,7 +1,14 @@
 import * as extension from "./extension";
 import * as utils from "./utils";
 import { basename } from "path";
-import * as vscode from "vscode";
+import {
+  CancellationToken,
+  DebugConfiguration,
+  DebugConfigurationProvider,
+  window,
+  ProviderResult,
+  WorkspaceFolder,
+} from "vscode";
 
 /**
  * Gets stringified settings to pass to the debug server.
@@ -11,41 +18,43 @@ export async function getDebugSettings() {
 }
 
 /**
- * Interacts with the user to create initial configurations.
+ * Interacts with the user to create a `roslaunch` or `rosrun` configuration.
  */
-export async function provideInitialConfigurations() {
-  const packages = utils.getPackages();
-
-  const command = await vscode.window.showQuickPick(["roslaunch", "rosrun"], { placeHolder: "Launch command" });
-  const packageName = await vscode.window.showQuickPick(packages.then(Object.keys), { placeHolder: "Package" });
-
-  let target: string;
-
-  if (packageName) {
-    let basenames = (files: string[]) => files.map(file => basename(file));
-
-    if (command === "roslaunch") {
-      const launches = utils.findPackageLaunchFiles(packageName).then(basenames);
-      target = await vscode.window.showQuickPick(launches, { placeHolder: "Launch file" });
-    } else {
-      const executables = utils.findPackageExecutables(packageName).then(basenames);
-      target = await vscode.window.showQuickPick(executables, { placeHolder: "Executable" });
-    }
-  } else {
-    target = await vscode.window.showInputBox({ placeHolder: "Target" });
+export class RosDebugConfigProvider implements DebugConfigurationProvider {
+  provideDebugConfigurations(folder: WorkspaceFolder | undefined, token?: CancellationToken) {
+    return [];
   }
 
-  return JSON.stringify({
-    configurations: [
-      {
-        command,
-        debugSettings: "${command:debugSettings}",
-        name: target,
-        package: packageName,
-        type: "ros",
-        target,
-      },
-    ],
-    version: "0.1.0",
-  }, undefined, 2);
+  async resolveDebugConfiguration(folder: WorkspaceFolder | undefined, config: DebugConfiguration, token?: CancellationToken) {
+    const packages = utils.getPackages();
+
+    const command = await window.showQuickPick(["roslaunch", "rosrun"], {  placeHolder: "Launch command" });
+    const packageName = await window.showQuickPick(packages.then(Object.keys), { placeHolder: "Package" });
+
+    let target: string;
+
+    if (packageName) {
+      let basenames = (files: string[]) => files.map(file => basename(file));
+
+      if (command === "roslaunch") {
+        const launches = utils.findPackageLaunchFiles(packageName).then(basenames);
+        target = await window.showQuickPick(launches, { placeHolder: "Launch file" });
+      } else {
+        const executables = utils.findPackageExecutables(packageName).then(basenames);
+        target = await window.showQuickPick(executables, { placeHolder: "Executable" });
+      }
+    } else {
+      target = await window.showInputBox({ placeHolder: "Target" });
+    }
+
+    config.type = "ros";
+    config.request = "launch";
+    config.command = command;
+    config.package = packageName;
+    config.target = target;
+    config.args = [];
+    config.debugSettings = "${command:debugSettings}";
+
+    return config;
+  }
 }